test(EntityProxyServer): focus codec service regression
